backend/requirements.txt
click==7.1.2
dnspython==2.0.0
email-validator==1.3.0
fastapi==0.61.1
h11==0.14.0
httptools==0.1.2
idna==3.4
motor==2.3.0
pydantic==1.7
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
six==1.16.0
starlette==0.13.6
uvicorn==0.12.2
uvloop==0.17.0
watchgod==0.6
websockets==8.1
